refactor: playlist, comment, gpt service interface 삭제후 service impl class 단일 클래스로 수정
